+++ src/subcommands/modpack/configure.rs
@@ -6,19 +6,19 @@ use dialoguer::Confirm;
 use libium::{config::structs::Modpack, file_picker::pick_folder};
 use std::path::PathBuf;
 
-pub async fn configure(
+pub fn configure(
     modpack: &mut Modpack,
     output_dir: Option<PathBuf>,
     install_overrides: Option<bool>,
 ) -> Result<()> {
     match output_dir {
         Some(output_dir) => {
-            check_output_directory(&output_dir).await?;
-            modpack.output_dir = output_dir.clone();
+            check_output_directory(&output_dir)?;
+            modpack.output_dir = output_dir;
         },
         None => {
-            if let Some(dir) = pick_folder(&modpack.output_dir, "Pick an output directory").await {
-                check_output_directory(&dir).await?;
+            if let Some(dir) = pick_folder(&modpack.output_dir, "Pick an output directory")? {
+                check_output_directory(&dir)?;
                 modpack.output_dir = dir;
             }
         },